pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal-session-recorder"
version = "2.0.0"
description = "Enterprise-grade Terminal Session Recorder for Pentesting & Security Audits"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
authors = [
    {name = "Md. Jony Hassain", email = "jonyhossan110@gmail.com"}
]
keywords = ["pentesting", "security", "terminal", "logging", "audit", "session-recording"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Logging",
]

dependencies = [
    "reportlab>=4.0.0",
    "Pillow>=10.0.0",
    "aiofiles>=23.0.0",
    "aiosqlite>=0.19.0",
    "psutil>=5.9.0",
    "rich>=13.0.0",
    "textual>=0.40.0",
    "asciinema>=2.3.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "flask>=3.0.0",
    "flask-socketio>=5.3.0",
    "requests>=2.31.0",
    "pygments>=2.16.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
]
cloud = [
    "boto3>=1.28.0",
    "google-cloud-storage>=2.10.0",
]
monitoring = [
    "mss>=9.0.1",
    "scapy>=2.5.0",
]
all = [
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    "boto3>=1.28.0",
    "mss>=9.0.1",
    "scapy>=2.5.0",
]

[project.urls]
Homepage = "https://github.com/jonyhossan110/terminal-session-recorder"
Repository = "https://github.com/jonyhossan110/terminal-session-recorder"
Issues = "https://github.com/jonyhossan110/terminal-session-recorder/issues"
Documentation = "https://github.com/jonyhossan110/terminal-session-recorder#readme"

[project.scripts]
tsr = "tsr.cli:main"
tsr-server = "tsr.server:main"
tsr-replay = "tsr.replay:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tsr*"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
